Created Map data structure
In order to implement the Life class, I needed a hash table
implementation of a map data structure. The one created obeys the
associative array abstraction, and also takes the key type, value type,
and hash functor as template parameters. It lacks a method to remove
key/value entries as it is not needed for the Life class.