Fixed bug when adding after resizing
Before, the resize occurs directly after failing to find a matching key
in the buckets array. The key-value pair is then added to the index
calculated previously, which has no relation to where it should hash to
in the new array. Thus, the code for resizing was put before searching
for the key.