Fixed linked list containing invalid buckets
When resizing the map, the linked list from the map copy was not copied
along with the rest of the members. This caused undefined behavior when
iterating over the map after a resize, because the bucket pointers
contained by the old linked list were invalidated. To fix this, the
linked list is copied from the map copy.