Refactored [] operator to use private method
The refactor introduces the helper method 'get' which does almost
everything the same as the previous implementation of the [] operator,
except that it returns a pointer to a bucket instead of the value by
reference. Then, it's simple to implement [] with this new method. The
reason for the change is to allow for some flexibility in how buckets
are tracked in future commits.