Optimized linked list update when resizing
Using the 'get' method introduced in a previous commit, when inserting a
key into the map copy, we get back the pointer to the bucket, which can
then be used to update the bucket pointer held by the linked list node
being iterated on. This helps avoid having to destroy the linked list
and copying the linked list of the map copy, and in fact, we also make
sure not to create a linked list in the map copy.