Changed using BucketType enum to using bool
The BucketType enum only had two states, 'EMPTY' or 'FILLED'. Thus, it
was simple to replace this with a bool named 'filled' that takes on
values of false, which implies that a bucket is empty, or true, which
implies that a bucket is filled. This should improve the memory usage
ever so slightly.