Created initial driver implementation
The driver program takes one argument, the path to a pattern file in RLE
format. At program launch, the driver prints to stdout the initial
configuration of the grid. Then, the user can enter a number to specify
how many generations to progress by, or a blank line to simply progress
by 1 generation. The grid configuration after the progression is printed
to stdout. The user can repeat this process as many times as they want
to. The program ends upon reaching EOF.