Created timer program
The timer program is used to time GoL simulations, as the name implies.
The timer program takes two arguments. The first one, PATTERN_FILENAME,
is the path to the RLE pattern file that is used to initialize the
automaton. The second one, ITERATIONS, is the number of times the
automaton will be progressed. The timer program prints to stdout some
information about the initial and final states of the automaton,
including the number of living cells (size) and the dimensions of the
grid (dims). Afterwards, it prints the total time it took to progress
the automaton.