Changed 'get' to default initialize value when inserting
Previously, when receiving a key that's not in the map as argument, the
Map methods 'get' and [] operator overload insert the key into
the map and return the corresponding uninitialized value. This commit
changes this behavior slightly, by first default initializing the value
before returning it. This change causes problems with value types that do
not have a default constructor, thus, only value types with default
constructors should be used with Map.