Optimized counter_increment_cell
The previous commit introduced a change that leads to the optimization
made in this commit. Specifically, the [] operator overload in Map now
default initializes the value when inserting a key. Thus, when inserting
a position into 'counter', the value is default initialized to 0. This
allows for skipping the check for whether a position is in 'counter'
with the 'contains' method, resulting in a non insignificant amount of
time being saved in the algorithm.