Changed Map internal list from linked list to vector
Previously, a linked list data structure was used to keep track of
pointers to all the filled buckets in the map. This was done to make
iteration on the order of the number of inserted keys in the map. The
linked list data structure was inefficient, however, probably due to the
pointer dereferences from following links. This commit changes the
list data structure to std::vector, which greatly improves the map
efficiency, according to tests run with the Life timer client.